🐛 fix:고정userid값 토큰에서 받아오도록 수정

+++ src/carts/carts.service.ts
@@ -7,12 +7,8 @@ import { calculateShippingFee } from '@src/shared/utils/shipping.util';
 export class CartsService {
   public constructor(private readonly prisma: PrismaService) {}
 
-  public async addToCart(cartId: string, productId: string): Promise<CartItem> {
-    const userId: string = 'xcl94l94lyb6dqceqi71r7z3'; // 임시 userId 설정
-
-    const cart = (await this.prisma.cart.findUnique({
-      where: { id: cartId },
-    })) as Cart;
+  public async addToCart(userId: string, cartId: string, productId: string): Promise<CartItem> {
+    const cart = await this.prisma.cart.findUnique({ where: { id: cartId } });
 
     if (!cart) {
       throw new NotFoundException('장바구니를 찾을 수 없습니다.');
@@ -30,21 +26,23 @@ export class CartsService {
       throw new NotFoundException('존재하지 않는 상품입니다.');
     }
 
-    let item = (await this.prisma.cartItem.findUnique({
+    let item = await this.prisma.cartItem.findUnique({
       where: { cartId_productId: { cartId, productId } },
-    })) as CartItem;
+    });
 
     if (!item) {
-      item = (await this.prisma.cartItem.create({
+      item = await this.prisma.cartItem.create({
         data: { cartId, productId, quantity: 1 },
-      })) as CartItem;
+      });
     }
 
     return item;
   }
 
-  //장바구니 전체조회
-  public async getCartItems(cartId: string): Promise<{
+  public async getCartItems(
+    userId: string,
+    cartId: string,
+  ): Promise<{
     items: CartItem[];
     totalAmount: number;
     shippingFee: number;
@@ -58,11 +56,10 @@ export class CartsService {
       },
     });
 
-    if (!cart) {
-      throw new NotFoundException('장바구니를 찾을 수 없습니다.');
+    if (!cart || cart.userId !== userId) {
+      throw new ForbiddenException('잘못된 장바구니 접근입니다.');
     }
 
-    // 총 금액 계산 (수량 * 상품 가격)
     const totalAmount = cart.cartItems.reduce((acc, item) => {
       return acc + item.quantity * item.product.price;
     }, 0);
@@ -76,14 +73,19 @@ export class CartsService {
     };
   }
 
-  //장바구니 수량 변경
-  public async updateCartItem(cartId: string, itemId: string, quantity: number): Promise<CartItem> {
+  public async updateCartItem(
+    userId: string,
+    cartId: string,
+    quantity: number,
+    itemId: string,
+  ): Promise<CartItem> {
     const item = await this.prisma.cartItem.findUnique({
       where: { id: itemId },
+      include: { cart: true },
     });
 
-    if (!item || item.cartId !== cartId) {
-      throw new NotFoundException('장바구니 항목을 찾을 수 없습니다.');
+    if (!item || item.cartId !== cartId || item.cart.userId !== userId) {
+      throw new ForbiddenException('잘못된 장바구니 접근입니다.');
     }
 
     return await this.prisma.cartItem.update({
@@ -92,14 +94,17 @@ export class CartsService {
     });
   }
 
-  //장바구니 상품 삭제
-  public async deleteCartItems(cartId: string, itemIds: string[]): Promise<{ message: string }> {
+  public async deleteCartItems(
+    userId: string,
+    cartId: string,
+    itemIds: string[],
+  ): Promise<{ message: string }> {
     const cart = await this.prisma.cart.findUnique({
       where: { id: cartId },
     });
 
-    if (!cart) {
-      throw new NotFoundException('장바구니를 찾을 수 없습니다.');
+    if (!cart || cart.userId !== userId) {
+      throw new ForbiddenException('잘못된 장바구니 접근입니다.');
     }
 
     await this.prisma.cartItem.deleteMany({
